Improve common artifact handle code
