+++ backend/src/main/java/ai/verta/modeldb/experimentRun/subtypes/DatasetHandler.java
@@ -1,9 +1,7 @@
 package ai.verta.modeldb.experimentRun.subtypes;
 
-import ai.verta.modeldb.common.exceptions.InternalErrorException;
 import ai.verta.modeldb.common.exceptions.ModelDBException;
 import ai.verta.modeldb.common.futures.FutureJdbi;
-import ai.verta.modeldb.common.subtypes.ArtifactHandlerBase;
 import ai.verta.modeldb.config.MDBConfig;
 import java.util.concurrent.Executor;
 
@@ -14,22 +12,8 @@ public class DatasetHandler extends ArtifactHandlerBase {
   }
 
   @Override
-  protected void setEntityIdReferenceColumn(String entityName) {
-    switch (entityName) {
-      case "ProjectEntity":
-        this.entityIdReferenceColumn = "project_id";
-        break;
-      case "ExperimentRunEntity":
-        this.entityIdReferenceColumn = "experiment_run_id";
-        break;
-      default:
-        throw new InternalErrorException("Invalid entity name: " + entityName);
-    }
-  }
-
-  @Override
   public void validateMaxArtifactsForTrial(int newArtifactsCount, int existingArtifactsCount)
       throws ModelDBException {
-    // No any logic for datasets
+    // No any trial restrictions are implemented for datasets
   }
 }
